requirements.txt
colorama==0.4.6
iniconfig==2.1.0
Nuitka==2.7.3
ordered-set==4.1.0
packaging==25.0
pluggy==1.6.0
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
pytest==8.3.5
PyYAML==6.0.2
shiboken6==6.9.0
zstandard==0.23.0
